Refactor projection forward pass (#334)
Refactors forward pass of projection to use tensor accessors and wrap kernel call in a class.
Also load cameras into shared memory which should be more efficient than having each thread load the same camera.

+++ fvdb/src/detail/ops/Ops.h
@@ -449,7 +449,7 @@ dispatchGaussianProjectionForward(const torch::Tensor &means,              // [N
                                   const torch::Tensor &scales,             // [N, 3]
                                   const torch::Tensor &worldToCamMatrices, // [C, 4, 4]
                                   const torch::Tensor &projectionMatrices, // [C, 3, 3]
-                                  const uint32_t imageWidth, const uint32_t imageHeight,
+                                  const int64_t imageWidth, const int64_t imageHeight,
                                   const float eps2d, const float nearPlane, const float farPlane,
                                   const float minRadius2d, const bool calcCompensations,
                                   const bool ortho);
